// travel-budget-spa/src/Interfaces/Enums.tsx
export enum StatementType {
    Income = 'INCOME',
    Expence = 'EXPENCE',
}

export enum StatementCategory {
    CampingSpot = 'CAMPING SPOT',
    Food = 'FOOD',
    Insurance = 'INSURANCE',
    PhoneAndInternet = 'PHONE + INTERNET',
    MediaAndEntertainment = 'MEDIA + ENTERTAINMENT',
    Income = 'INCOME',
}